BehandlingMetode på behandlingTilstand
