+++ integrasjontjenester/datavarehus/src/main/java/no/nav/foreldrepenger/tilbakekreving/datavarehus/saksstatistikk/BehandlingTilstandTjeneste.java
@@ -1,6 +1,8 @@
 package no.nav.foreldrepenger.tilbakekreving.datavarehus.saksstatistikk;
 
+import java.time.LocalDateTime;
 import java.time.OffsetDateTime;
+import java.time.ZoneId;
 import java.time.ZoneOffset;
 import java.util.Optional;
 
@@ -12,6 +14,8 @@ import no.nav.foreldrepenger.tilbakekreving.behandlingslager.behandling.Behandli
 import no.nav.foreldrepenger.tilbakekreving.behandlingslager.behandling.BehandlingResultatType;
 import no.nav.foreldrepenger.tilbakekreving.behandlingslager.behandling.Behandlingsresultat;
 import no.nav.foreldrepenger.tilbakekreving.behandlingslager.behandling.BehandlingÅrsak;
+import no.nav.foreldrepenger.tilbakekreving.behandlingslager.behandling.aksjonspunkt.Aksjonspunkt;
+import no.nav.foreldrepenger.tilbakekreving.behandlingslager.behandling.aksjonspunkt.AksjonspunktDefinisjon;
 import no.nav.foreldrepenger.tilbakekreving.behandlingslager.behandling.aksjonspunkt.Venteårsak;
 import no.nav.foreldrepenger.tilbakekreving.behandlingslager.behandling.ekstern.EksternBehandling;
 import no.nav.foreldrepenger.tilbakekreving.behandlingslager.behandling.repository.BehandlingRepository;
@@ -23,6 +27,7 @@ import no.nav.foreldrepenger.tilbakekreving.datavarehus.saksstatistikk.mapping.B
 import no.nav.foreldrepenger.tilbakekreving.datavarehus.saksstatistikk.mapping.BehandlingTypeMapper;
 import no.nav.foreldrepenger.tilbakekreving.datavarehus.saksstatistikk.mapping.BehandlingÅrsakMapper;
 import no.nav.foreldrepenger.tilbakekreving.datavarehus.saksstatistikk.mapping.YtelseTypeMapper;
+import no.nav.foreldrepenger.tilbakekreving.kontrakter.felles.BehandlingMetode;
 import no.nav.foreldrepenger.tilbakekreving.kontrakter.sakshendelse.BehandlingTilstand;
 import no.nav.foreldrepenger.tilbakekreving.kontrakter.sakshendelse.Periode;
 
@@ -72,11 +77,15 @@ public class BehandlingTilstandTjeneste {
         tilstand.setBehandlingType(BehandlingTypeMapper.getBehandlingType(behandling.getType()));
         tilstand.setBehandlingStatus(BehandlingStatusMapper.getBehandlingStatus(behandling.getStatus()));
         tilstand.setBehandlingResultat(BehandlingResultatTypeMapper.getBehandlingResultatType(behandlingResultatType));
+        tilstand.setBehandlingMetode(utledBehandlingMetode(behandling));
         tilstand.setBehandlendeEnhetKode(behandling.getBehandlendeEnhetId());
         tilstand.setAnsvarligBeslutter(behandling.getAnsvarligBeslutter());
         tilstand.setAnsvarligSaksbehandler(behandling.getAnsvarligSaksbehandler());
         tilstand.setErBehandlingManueltOpprettet(erSaksbehandler(behandling.getOpprettetAv()));
-        tilstand.setFunksjonellTid(OffsetDateTime.now(ZoneOffset.UTC));
+        tilstand.setOpprettetAv(behandling.getOpprettetAv());
+        tilstand.setFunksjonellTid(behandling.erAvsluttet() ? tilOffsetDateTime(behandling.getEndretTidspunkt()) : OffsetDateTime.now(ZoneOffset.UTC));
+        tilstand.setRegistrertTid(tilOffsetDateTime(behandling.getOpprettetTidspunkt()));
+        tilstand.setFerdigBehandletTid(behandling.erAvsluttet() ? tilOffsetDateTime(behandling.getEndretTidspunkt()) : null);
         tilstand.setVenterPåBruker(venterPåBruker);
         tilstand.setVenterPåØkonomi(venterPåØkonomi);
         forrigeBehandling.ifPresent(forrige -> tilstand.setForrigeBehandling(forrige.getUuid()));
@@ -101,6 +110,30 @@ public class BehandlingTilstandTjeneste {
             eksternBehandling = eksternBehandlingRepository.hentForSisteAktivertInternId(behandlingId);
         }
         return eksternBehandling;
+    }
+
+    private static OffsetDateTime tilOffsetDateTime(LocalDateTime tidspunkt) {
+        return OffsetDateTime.ofInstant(tidspunkt.atZone(ZoneId.systemDefault()).toInstant(), ZoneOffset.UTC);
+    }
+
+    private static BehandlingMetode utledBehandlingMetode(Behandling behandling) {
+        if (!behandling.erSaksbehandlingAvsluttet()) {
+            return null;
+        }
+        if (behandling.getAksjonspunktMedDefinisjonOptional(AksjonspunktDefinisjon.FATTE_VEDTAK).filter(Aksjonspunkt::erUtført).isPresent()) {
+            return BehandlingMetode.TOTRINN;
+        }
+        if (behandling.getAksjonspunkter().stream().filter(ap -> !ap.erAutopunkt()).anyMatch(BehandlingTilstandTjeneste::harSaksbehandlerVurdertAksjonspunkt)) {
+            return BehandlingMetode.MANUELL;
+        }
+        if (behandling.getAksjonspunkter().stream().filter(Aksjonspunkt::erAutopunkt).anyMatch(BehandlingTilstandTjeneste::harSaksbehandlerVurdertAksjonspunkt)) {
+            return BehandlingMetode.INNHENTING;
+        }
+        return BehandlingMetode.AUTOMATISK;
+    }
+
+    private static boolean harSaksbehandlerVurdertAksjonspunkt(Aksjonspunkt aksjonspunkt) {
+        return aksjonspunkt.erUtført() || erSaksbehandler(aksjonspunkt.getEndretAv()) || erSaksbehandler(aksjonspunkt.getOpprettetAv());
     }
 
     private static boolean erSaksbehandler(String s) {
