+++ integrasjontjenester/datavarehus/src/main/java/no/nav/foreldrepenger/tilbakekreving/datavarehus/saksstatistikk/aiven/AivenVedtakKafkaProducer.java
@@ -1,12 +1,14 @@
 package no.nav.foreldrepenger.tilbakekreving.datavarehus.saksstatistikk.aiven;
 
-import jakarta.enterprise.context.ApplicationScoped;
-import jakarta.inject.Inject;
+import java.time.OffsetDateTime;
+import java.time.ZoneOffset;
 
 import org.apache.kafka.clients.producer.ProducerRecord;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
+import jakarta.enterprise.context.ApplicationScoped;
+import jakarta.inject.Inject;
 import no.nav.foreldrepenger.konfig.KonfigVerdi;
 import no.nav.foreldrepenger.tilbakekreving.datavarehus.saksstatistikk.mapping.VedtakOppsummeringMapper;
 import no.nav.foreldrepenger.tilbakekreving.integrasjon.kafka.AivenMeldingProducer;
@@ -27,6 +29,8 @@ public class AivenVedtakKafkaProducer extends AivenMeldingProducer {
     }
 
     public void sendMelding(VedtakOppsummering vedtakOppsummering) {
+        vedtakOppsummering.setTekniskTid(OffsetDateTime.now(ZoneOffset.UTC)); //tidspunkt for sending
+
         var nøkkel = vedtakOppsummering.getBehandlingUuid().toString();
         var verdi = VedtakOppsummeringMapper.tilJsonString(vedtakOppsummering);
         var melding = new ProducerRecord<>(getTopic(), nøkkel, verdi);
