+++ integrasjontjenester/datavarehus/src/test/java/no/nav/foreldrepenger/tilbakekreving/datavarehus/saksstatistikk/BehandlingTilstandTjenesteTest.java
@@ -5,6 +5,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 
 import java.time.LocalDate;
 import java.time.OffsetDateTime;
+import java.time.ZoneOffset;
 import java.util.UUID;
 
 import org.junit.jupiter.api.BeforeEach;
@@ -16,7 +17,10 @@ import no.nav.foreldrepenger.tilbakekreving.behandling.impl.BehandlingTjeneste;
 import no.nav.foreldrepenger.tilbakekreving.behandlingslager.aktør.OrganisasjonsEnhet;
 import no.nav.foreldrepenger.tilbakekreving.behandlingslager.behandling.Behandling;
 import no.nav.foreldrepenger.tilbakekreving.behandlingslager.behandling.BehandlingResultatType;
+import no.nav.foreldrepenger.tilbakekreving.behandlingslager.behandling.BehandlingStegType;
 import no.nav.foreldrepenger.tilbakekreving.behandlingslager.behandling.Behandlingsresultat;
+import no.nav.foreldrepenger.tilbakekreving.behandlingslager.behandling.aksjonspunkt.AksjonspunktDefinisjon;
+import no.nav.foreldrepenger.tilbakekreving.behandlingslager.behandling.aksjonspunkt.AksjonspunktKontrollRepository;
 import no.nav.foreldrepenger.tilbakekreving.behandlingslager.behandling.aksjonspunkt.Venteårsak;
 import no.nav.foreldrepenger.tilbakekreving.behandlingslager.behandling.ekstern.EksternBehandling;
 import no.nav.foreldrepenger.tilbakekreving.behandlingslager.behandling.repository.BehandlingRepositoryProvider;
@@ -24,6 +28,7 @@ import no.nav.foreldrepenger.tilbakekreving.behandlingslager.behandling.reposito
 import no.nav.foreldrepenger.tilbakekreving.behandlingslager.testutilities.kodeverk.ScenarioSimple;
 import no.nav.foreldrepenger.tilbakekreving.dbstoette.CdiDbAwareTest;
 import no.nav.foreldrepenger.tilbakekreving.domene.typer.Henvisning;
+import no.nav.foreldrepenger.tilbakekreving.kontrakter.felles.BehandlingMetode;
 import no.nav.foreldrepenger.tilbakekreving.kontrakter.felles.BehandlingResultat;
 import no.nav.foreldrepenger.tilbakekreving.kontrakter.felles.BehandlingStatus;
 import no.nav.foreldrepenger.tilbakekreving.kontrakter.felles.BehandlingType;
@@ -69,6 +74,9 @@ class BehandlingTilstandTjenesteTest {
         assertThat(tilstand.erBehandlingManueltOpprettet()).isFalse();
         assertThat(tilstand.getFunksjonellTid()).isBetween(OffsetDateTime.now().minusMinutes(1), OffsetDateTime.now());
         assertThat(tilstand.getTekniskTid()).isNull();
+        assertThat(tilstand.getRegistrertTid()).isBetween(OffsetDateTime.now().minusMinutes(1), OffsetDateTime.now());
+        assertThat(tilstand.getFerdigBehandletTid()).isNull();
+        assertThat(tilstand.getOpprettetAv()).isEqualTo("VL");
         assertThat(tilstand.getAnsvarligBeslutter()).isNull();
         assertThat(tilstand.getAnsvarligSaksbehandler()).isNull();
         assertThat(tilstand.getBehandlendeEnhetKode()).isNull();
@@ -76,6 +84,13 @@ class BehandlingTilstandTjenesteTest {
 
     @Test
     void skal_utlede_behandlingtilstand_for_fattet_behandling() {
+        var scenario = ScenarioSimple.simple();
+        scenario.leggTilAksjonspunkt(AksjonspunktDefinisjon.FATTE_VEDTAK, BehandlingStegType.FATTE_VEDTAK);
+        behandling = scenario.lagre(behandlingRepositoryProvider);
+        behandling.getÅpneAksjonspunkter().forEach(ap -> new AksjonspunktKontrollRepository().setTilUtført(ap));
+        behandlingRepositoryProvider.getBehandlingRepository().lagre(behandling, behandlingRepositoryProvider.getBehandlingRepository().taSkriveLås(behandling));
+        var eksternBehandling = new EksternBehandling(behandling, Henvisning.fraEksternBehandlingId(1l), EKSTERN_UUID);
+        behandlingRepositoryProvider.getEksternBehandlingRepository().lagre(eksternBehandling);
         behandlingresultatRepository.lagre(Behandlingsresultat.builder()
                 .medBehandling(behandling)
                 .medBehandlingResultatType(BehandlingResultatType.FULL_TILBAKEBETALING).build());
@@ -97,11 +112,15 @@ class BehandlingTilstandTjenesteTest {
         assertThat(tilstand.getBehandlingType()).isEqualTo(BehandlingType.TILBAKEKREVING);
         assertThat(tilstand.getBehandlingStatus()).isEqualTo(BehandlingStatus.AVSLUTTET);
         assertThat(tilstand.getBehandlingResultat()).isEqualTo(BehandlingResultat.FULL_TILBAKEBETALING);
+        assertThat(tilstand.getBehandlingMetode()).isEqualTo(BehandlingMetode.TOTRINN);
         assertThat(tilstand.venterPåBruker()).isFalse();
         assertThat(tilstand.venterPåØkonomi()).isFalse();
         assertThat(tilstand.erBehandlingManueltOpprettet()).isFalse();
         assertThat(tilstand.getFunksjonellTid()).isBetween(OffsetDateTime.now().minusMinutes(1), OffsetDateTime.now());
         assertThat(tilstand.getTekniskTid()).isNull();
+        assertThat(tilstand.getRegistrertTid()).isBetween(OffsetDateTime.now().minusMinutes(1), OffsetDateTime.now());
+        assertThat(tilstand.getFerdigBehandletTid()).isBetween(OffsetDateTime.now().minusMinutes(1), OffsetDateTime.now());
+        assertThat(tilstand.getOpprettetAv()).isEqualTo("VL");
         assertThat(tilstand.getAnsvarligBeslutter()).isEqualTo("Z111112");
         assertThat(tilstand.getAnsvarligSaksbehandler()).isEqualTo("Z111111");
         assertThat(tilstand.getBehandlendeEnhetKode()).isEqualTo("1234");
@@ -125,5 +144,8 @@ class BehandlingTilstandTjenesteTest {
         assertThat(tilstand.venterPåBruker()).isTrue();
         assertThat(tilstand.venterPåØkonomi()).isFalse();
         assertThat(tilstand.getFunksjonellTid()).isBetween(OffsetDateTime.now().minusMinutes(1), OffsetDateTime.now());
+        assertThat(tilstand.getRegistrertTid()).isBetween(OffsetDateTime.now().minusMinutes(1), OffsetDateTime.now());
+        assertThat(tilstand.getFerdigBehandletTid()).isNull();
+        assertThat(tilstand.getOpprettetAv()).isEqualTo("VL");
     }
 }
