+++ integrasjontjenester/datavarehus/src/test/java/no/nav/foreldrepenger/tilbakekreving/datavarehus/saksstatistikk/VedtakOppsummeringTjenesteTest.java
@@ -18,12 +18,9 @@ import no.nav.foreldrepenger.tilbakekreving.behandling.beregning.TilbakekrevingB
 import no.nav.foreldrepenger.tilbakekreving.behandling.impl.KravgrunnlagBeregningTjeneste;
 import no.nav.foreldrepenger.tilbakekreving.behandlingslager.behandling.Behandling;
 import no.nav.foreldrepenger.tilbakekreving.behandlingslager.behandling.BehandlingResultatType;
-import no.nav.foreldrepenger.tilbakekreving.behandlingslager.behandling.BehandlingStegType;
 import no.nav.foreldrepenger.tilbakekreving.behandlingslager.behandling.Behandlingsresultat;
 import no.nav.foreldrepenger.tilbakekreving.behandlingslager.behandling.ForeldelseVurderingType;
 import no.nav.foreldrepenger.tilbakekreving.behandlingslager.behandling.KlasseKode;
-import no.nav.foreldrepenger.tilbakekreving.behandlingslager.behandling.aksjonspunkt.AksjonspunktDefinisjon;
-import no.nav.foreldrepenger.tilbakekreving.behandlingslager.behandling.aksjonspunkt.AksjonspunktKontrollRepository;
 import no.nav.foreldrepenger.tilbakekreving.behandlingslager.behandling.ekstern.EksternBehandling;
 import no.nav.foreldrepenger.tilbakekreving.behandlingslager.behandling.repository.BehandlingRepositoryProvider;
 import no.nav.foreldrepenger.tilbakekreving.behandlingslager.behandling.repository.EksternBehandlingRepository;
@@ -61,7 +58,6 @@ import no.nav.foreldrepenger.tilbakekreving.grunnlag.KravgrunnlagRepository;
 import no.nav.foreldrepenger.tilbakekreving.grunnlag.kodeverk.GjelderType;
 import no.nav.foreldrepenger.tilbakekreving.grunnlag.kodeverk.KlasseType;
 import no.nav.foreldrepenger.tilbakekreving.grunnlag.kodeverk.KravStatusKode;
-import no.nav.foreldrepenger.tilbakekreving.kontrakter.felles.BehandlingMetode;
 import no.nav.foreldrepenger.tilbakekreving.kontrakter.felles.BehandlingType;
 import no.nav.foreldrepenger.tilbakekreving.kontrakter.felles.YtelseType;
 import no.nav.foreldrepenger.tilbakekreving.kontrakter.vedtak.UtvidetVilkårResultat;
@@ -104,14 +100,10 @@ class VedtakOppsummeringTjenesteTest {
         vedtakOppsummeringTjeneste = new VedtakOppsummeringTjeneste(repositoryProvider, beregningsresultatTjeneste);
 
         entityManager.setFlushMode(FlushModeType.AUTO);
-        var scenario = ScenarioSimple.simple();
-        scenario.leggTilAksjonspunkt(AksjonspunktDefinisjon.FATTE_VEDTAK, BehandlingStegType.FATTE_VEDTAK);
-        behandling = scenario.lagre(repositoryProvider);
-        behandling.getÅpneAksjonspunkter().forEach(ap -> new AksjonspunktKontrollRepository().setTilUtført(ap));
+        behandling = ScenarioSimple.simple().lagre(repositoryProvider);
         behandling.setAnsvarligSaksbehandler(ANSVARLIG_SAKSBEHANDLER);
         behandling.setAnsvarligBeslutter(ANSVARLIG_BESLUTTER);
         behandling.setBehandlendeEnhetId("8020");
-        behandling.avsluttBehandling();
         entityManager.persist(behandling);
 
         behandlingId = behandling.getId();
@@ -188,10 +180,6 @@ class VedtakOppsummeringTjenesteTest {
         assertThat(vedtakOppsummering.getVedtakFattetTid()).isNotNull();
         assertThat(vedtakOppsummering.getYtelseType()).isEqualByComparingTo(YtelseType.FP);
         assertThat(vedtakOppsummering.getForrigeBehandling()).isNull();
-        assertThat(vedtakOppsummering.getBehandlingMetode()).isEqualTo(BehandlingMetode.TOTRINN);
-        assertThat(vedtakOppsummering.getTekniskTid()).isNotNull();
-        assertThat(vedtakOppsummering.getFerdigBehandletTid()).isNotNull();
-        assertThat(vedtakOppsummering.getOpprettetAv()).isEqualTo("VL");
     }
 
     private VedtakPeriode fellesAssertVedtakPeriode(List<VedtakPeriode> vedtakPerioder) {
