+++ kontrakter/src/main/java/no/nav/foreldrepenger/tilbakekreving/kontrakter/vedtak/VedtakOppsummering.java
@@ -6,7 +6,6 @@ import java.util.UUID;
 
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Size;
-import no.nav.foreldrepenger.tilbakekreving.kontrakter.felles.BehandlingMetode;
 import no.nav.foreldrepenger.tilbakekreving.kontrakter.felles.BehandlingType;
 import no.nav.foreldrepenger.tilbakekreving.kontrakter.felles.YtelseType;
 
@@ -24,9 +23,7 @@ public class VedtakOppsummering {
     private UUID referertFagsakBehandlingUuid;
     @NotNull
     private BehandlingType behandlingType;
-    private BehandlingMetode behandlingMetode;
     private boolean erBehandlingManueltOpprettet;
-    private String opprettetAv;
     @NotNull
     private String behandlendeEnhetKode;
     @NotNull
@@ -37,7 +34,6 @@ public class VedtakOppsummering {
     private OffsetDateTime behandlingOpprettetTid;
     @NotNull
     private OffsetDateTime vedtakFattetTid;
-    private OffsetDateTime ferdigBehandletTid;
     @NotNull
     @Size(min = 1, max = 100)
     private List<VedtakPeriode> perioder;
@@ -98,28 +94,12 @@ public class VedtakOppsummering {
         this.behandlingType = behandlingType;
     }
 
-    public BehandlingMetode getBehandlingMetode() {
-        return behandlingMetode;
-    }
-
-    public void setBehandlingMetode(BehandlingMetode behandlingMetode) {
-        this.behandlingMetode = behandlingMetode;
-    }
-
     public boolean isErBehandlingManueltOpprettet() {
         return erBehandlingManueltOpprettet;
     }
 
     public void setErBehandlingManueltOpprettet(boolean erBehandlingManueltOpprettet) {
         this.erBehandlingManueltOpprettet = erBehandlingManueltOpprettet;
-    }
-
-    public String getOpprettetAv() {
-        return opprettetAv;
-    }
-
-    public void setOpprettetAv(String opprettetAv) {
-        this.opprettetAv = opprettetAv;
     }
 
     public String getBehandlendeEnhetKode() {
@@ -160,14 +140,6 @@ public class VedtakOppsummering {
 
     public void setVedtakFattetTid(OffsetDateTime vedtakFattetTid) {
         this.vedtakFattetTid = vedtakFattetTid;
-    }
-
-    public OffsetDateTime getFerdigBehandletTid() {
-        return ferdigBehandletTid;
-    }
-
-    public void setFerdigBehandletTid(OffsetDateTime ferdigBehandletTid) {
-        this.ferdigBehandletTid = ferdigBehandletTid;
     }
 
     public List<VedtakPeriode> getPerioder() {
